ref(core): consolidate radio component files and remove chonk naming
Moves all content from radio.chonk.tsx into radio.tsx. Removes "chonk"
prefix from chonkRadioStyles function.
